gsm: use different brainstem for acroname USB HUB in rhel7

+++ prepare/envsetup.sh
@@ -568,9 +568,17 @@ install_usb_hub_driver_el () {
     yum install -y libffi-devel python36-devel
     pushd tmp/usb_hub
         # The module brainstem is already stored in project NetworkManager-ci.
-        tar xf brainstem_dev_kit_ubuntu_lts_18.04_no_qt_x86_64_1.tgz
-        cd development/python/
-        python -m pip install brainstem-2.7.0-py2.py3-none-any.whl; local rc=$?
+        if grep -q -e 'Enterprise Linux .*release 7' -e 'CentOS Linux release 7' /etc/redhat-release; then
+            # Compatible with RHEL7
+            tar xf brainstem_dev_kit_ubuntu_lts_14.04_x86_64.tgz
+            cd development/python/
+            python -m pip install brainstem-2.7.1-py2.py3-none-any.whl; local rc=$?
+        else
+            # And with RHEL8
+            tar xf brainstem_dev_kit_ubuntu_lts_18.04_no_qt_x86_64_1.tgz
+            cd development/python/
+            python -m pip install brainstem-2.7.0-py2.py3-none-any.whl; local rc=$?
+        fi
     popd
     return $rc
 }
